refactor(mqtt): Modify unreasonable code
* Modify header file ordering
* Modify mqtt.* to mqtt_*.* to avoid ide parsing exception
* Modify part of the code according to the pr revision comments

+++ include/neuron/json/neu_json_mqtt.h
@@ -43,7 +43,7 @@ typedef struct {
 
 typedef struct {
     char *   version;
-    uint32_t timpstamp;
+    uint64_t timpstamp;
 } neu_json_states_head_t;
 
 typedef struct {
